# backend/openapi.yaml
openapi: 3.0.3
info:
  title: BANKERCHANGER API
  version: 0.2.0
  description: REST API for the BANKERCHANGER decentralized boxing prediction market.
servers:
  - url: http://localhost:3001
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    oracleKey:
      type: apiKey
      in: header
      name: X-Oracle-Key
  schemas:
    Error:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string }
    ErrorResponse400:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string, description: "Bad request. Invalid query parameters, request body, or format." }
        code: { type: string, example: "INVALID_REQUEST" }
      required: [success, message]
...
    ErrorResponse429:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string, description: "Rate limit exceeded. Check Retry-After header for reset time." }
        code: { type: string, example: "RATE_LIMIT_EXCEEDED" }
        retryAfter: { type: integer, description: "Seconds until the rate limit window resets." }
      required: [success, message]
    ErrorResponse500:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string, description: "Internal server error. Please try again later." }
        code: { type: string, example: "INTERNAL_SERVER_ERROR" }
        requestId: { type: string, description: "Unique identifier for this error instance for debugging." }
      required: [success, message]
...
    Market:
      type: object
      properties:
        id:               { type: integer }
        market_id:        { type: string }
        contract_address: { type: string }
        match_id:         { type: string }
        fighter_a:        { type: string }
        fighter_b:        { type: string }
        weight_class:     { type: string }
        title_fight:      { type: boolean }
        venue:            { type: string }
        scheduled_at:     { type: string, format: date-time }
        status:
          type: string
          enum: [open, locked, resolved, cancelled, disputed]
        outcome:
          type: string
          nullable: true
          enum: [fighter_a, fighter_b, draw, no_contest]
        pool_a:           { type: string }
        pool_b:           { type: string }
        pool_draw:        { type: string }
        total_pool:       { type: string }
        fee_bps:          { type: integer }
        resolved_at:      { type: string, format: date-time, nullable: true }
        oracle_used:      { type: string, nullable: true, enum: [primary, fallback, admin] }
        created_at:       { type: string, format: date-time }
        updated_at:       { type: string, format: date-time }
        ledger_sequence:  { type: integer }
...
paths:
...
  /api/markets:
    get:
      summary: List markets
      operationId: listMarkets
      tags: [Markets]
      x-rateLimit-windowMs: 60000
      x-rateLimit-max: 60
      x-rateLimit-keyBy: ip
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, locked, resolved, cancelled, disputed]
        - name: weight_class
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Paginated list of markets
          content:
            application/json:
              schema:
                type: object
                properties:
                  markets:
                    type: array
                    items: { $ref: '#/components/schemas/Market' }
                  total: { type: integer }
                  page:  { type: integer }
                  limit: { type: integer }
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse400' }
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse429' }
        '500':
          description: Internal server error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse500' }
  /api/markets/{market_id}:
    get:
      summary: Get market by ID
      operationId: getMarket
      tags: [Markets]
      parameters:
        - name: market_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Market detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Market' }
        '400':
          description: Invalid market ID format
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse400' }
        '404':
          description: Market not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse429' }
        '500':
          description: Internal server error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse500' }
...